added setValue method
setValue method is needed when value of NumberSelector is supposed to be set from outside. Currently it is requierd to create new instance of NumberSelector each time value is changed from outside.

I also added check for min max range, so that if value is set from outside of acceptable range, it will be ignored.